connecting searchbar to table WIP
