+++ components/Venues/Modal.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="modal-backdrop">
+    <div class="modal">
+      <div class="modal-content">
+        <slot />
+        <button type="button" class="btn-close" @click="handleClose">X</button>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script setup>
+const handleClose = () => {
+  console.log("yes");
+};
+</script>
