+++ pages/manage-venue.vue
@@ -1,23 +1,26 @@
 <!-- eslint-disable vue/attribute-hyphenation -->
 <template>
   <Header titleText="Venues" />
-  <SearchBar />
+  <SearchBar @handle-filter="filterVenues" />
   <div class="flex content-center justify-center">
     <Table :config="config" :venues="venues" />
   </div>
   <Button
-      button-text="Add Venue"
-      button-color="bg-gold"
-      @click="(show =! show)"
-    />
+    button-text="Add Venue"
+    button-color="bg-gold"
+    @click="show = !show"
+  />
   <Modal v-show="show">
     <h1>Add Venue</h1>
   </Modal>
 </template>
 <script setup>
+import { ref } from "vue";
+import Modal from "../components/Venues/Modal.vue";
 
-import { ref } from 'vue';
-import Modal from '../components/Modal.vue';
+const filterVenues = (value) => {
+  venues.filter(value);
+};
 
 const venues = [
   {
